# enhancemanet: reduce calculation calls