@sanposhiho 2023-04-16
@lowang-bh 
Sorry, I've just realized this change is the same as https://github.com/kubernetes/kubernetes/pull/117208 which was submitted a bit earlier than this.