@lowang-bh 2023-04-18
> @lowang-bh Sorry, I've just realized this change is the same as #117208 which was submitted a bit earlier than this.

Yeah, I see. I find my pr has less function calls than that one.
And, I take a benchmark on that modify, here is the result.

```sh
Benchmark10kPods-8   	   52952	     22429 ns/op	       0 B/op	       0 allocs/op
Benchmark10kPods-8   	   47960	     24871 ns/op	       0 B/op	       0 allocs/op
Benchmark100kPods-8   	    1060	   1126838 ns/op	       0 B/op	       0 allocs/op
Benchmark100kPods-8   	    1044	   1187247 ns/op	       0 B/op	       0 allocs/op

```